% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcm.R
\docType{package}
\name{fcm}
\alias{fcm}
\alias{fcm-package}
\title{Fuzzy Cognitive Maps (FCMs) Inference}
\description{
Estimates the Inference of a Fuzzy Cognitive Map. FCM is proven to be capable of causal inference and is applicable to complex decision problems where numerous interlinked dependent variables influence one another. The fcm package includes the fcm.infer function which returns a dataframe of the concepts' values of each state after the inference procedure.
}
